+++ src/ram.c
@@ -6,8 +6,8 @@
 
 SDL_Surface *g_screen;
 int g_scene;
-long gameflag[1024];
-long gameflag2[1024];
+uint32_t gameflag[1024];
+uint32_t gameflag2[1024];
 
 long g_DebugDisp;
 long g_DebugStageProduction;
